feat/AB#81960_ABC-in-charts-predefined-filters-when-using-reference-data-wont-work refactor: create get-reference-metadata file with reference data fields to meta function refactor: place that shared function in aggregation builder and predefined filters components
